render action panel
