+++ code/cell.modules/pkg.sys/ui.harness/src/common/index.ts
@@ -3,6 +3,7 @@ import * as constants from './constants';
 
 export { t, constants };
 export * from './libs';
+export * from './util';
 
 export const COLORS = constants.COLORS;
 export const DEFAULT = constants.DEFAULT;
