+++ code/cell.modules/pkg.sys/ui.harness/src/common/types.ts
@@ -1,5 +1,7 @@
 export * from '@platform/types';
 export * from '@platform/cell.types/lib/types.Module';
-export * from '@platform/state.types';
+
+export { IStateObject, IStateObjectWritable } from '@platform/state.types';
+export { CssValue } from '@platform/css';
 
 export * from '../types';
