+++ code/cell.modules/pkg.sys/ui.harness/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './types.Host';
-export * from './types.Model';
-export * from './types.Model.builder';
+export * from './types.ActionPanel';
+export * from './types.ActionPanel.model';
+export * from './types.ActionPanel.builder';
