+++ code/cell.modules/pkg.sys/ui.harness/src/types/types.ActionPanel.ts
@@ -1,0 +1,6 @@
+import { t } from './common';
+
+export type ActionPanelProps = {
+  showTitle?: boolean;
+  style?: t.CssValue;
+};
